#39 レース結果登録済みの場合に馬券購入時の払い戻し金額を計算するよう修正
