+++ source/app/UseCases/RaceResult/StoreAction.php
@@ -6,6 +6,7 @@ use App\Models\Race;
 use App\Models\RacePayout;
 use App\Models\RacePayoutHorse;
 use App\Models\TicketPurchase;
+use App\UseCases\TicketPurchase\CalculatePayoutAmountAction;
 use Illuminate\Support\Facades\DB;
 
 /**
@@ -41,17 +42,9 @@ class StoreAction
         'umaren', 'umatan', 'sanrenpuku', 'sanrentan',
     ];
 
-    /** 券種ごとの指定頭数（照合対象の馬番数） */
-    private const TICKET_TYPE_HORSE_COUNT = [
-        'tansho' => 1,
-        'fukusho' => 1,
-        'wakuren' => 2,
-        'umaren' => 2,
-        'umatan' => 2,
-        'wide' => 2,
-        'sanrenpuku' => 3,
-        'sanrentan' => 3,
-    ];
+    public function __construct(
+        private readonly CalculatePayoutAmountAction $calculatePayoutAmountAction,
+    ) {}
 
     /**
      * @param  array{text: string}  $data
@@ -102,39 +95,9 @@ class StoreAction
 
     /**
      * 対象レースの購入馬券（投稿ユーザーのみ）を取得し、レース結果と照合して payout_amount を更新する。
-     *
-     * 各 TicketPurchase の selections を buy_type に応じて組み合わせに展開し、
-     * 同じレースの race_payouts（ticket_type ごとの払い戻し行）の馬番配列と照合する。
-     * ヒットした組み合わせの JRA 払戻額（100円あたり）を合算し、購入金額に応じてスケールする。
-     * 購入金額（amount）が null の場合は更新しない。
-     *
-     * 計算式: ヒット組み合わせの払戻合計 × (amount / (組み合わせ数 × 100))
      */
     private function updateTicketPurchasesPayoutAmount(int $raceId, int $userId): void
     {
-        /** @var \Illuminate\Database\Eloquent\Collection<int, RacePayout> $payouts */
-        $payouts = RacePayout::with(['ticketType', 'racePayoutHorses' => function ($query): void {
-            $query->orderBy('sort_order');
-        }])->where('race_id', $raceId)->get();
-
-        /** @var array<string, array<int, array{amount: int, horse_numbers: array<int, int>}>> $payoutsByType */
-        $payoutsByType = [];
-        foreach ($payouts as $payout) {
-            $typeName = $payout->ticketType->name;
-            $horseNumbers = $payout->racePayoutHorses
-                ->sortBy('sort_order')
-                ->pluck('horse_number')
-                ->map(fn ($n) => (int) $n)
-                ->values()
-                ->all();
-
-            $payoutsByType[$typeName] ??= [];
-            $payoutsByType[$typeName][] = [
-                'amount' => (int) $payout->payout_amount,
-                'horse_numbers' => $horseNumbers,
-            ];
-        }
-
         /** @var \Illuminate\Database\Eloquent\Collection<int, TicketPurchase> $purchases */
         $purchases = TicketPurchase::with(['ticketType', 'buyType'])
             ->where('race_id', $raceId)
@@ -142,321 +105,12 @@ class StoreAction
             ->get();
 
         foreach ($purchases as $purchase) {
-            $ticketTypeName = $purchase->ticketType->name;
-            $buyTypeName = $purchase->buyType->name;
-            $selections = $purchase->selections ?? [];
-
-            $horseCount = self::TICKET_TYPE_HORSE_COUNT[$ticketTypeName] ?? null;
-            if ($horseCount === null) {
-                continue;
-            }
-
-            $isOrdered = in_array($ticketTypeName, self::ORDERED_TYPES, true);
-            $combinations = $this->expandSelections($buyTypeName, $selections, $horseCount, $isOrdered);
-
-            if ($combinations === []) {
-                continue;
-            }
-
-            $totalPayout = $this->sumMatchedPayouts(
-                $combinations,
-                $payoutsByType[$ticketTypeName] ?? [],
-                $isOrdered,
-            );
-
-            if ($totalPayout > 0 && $purchase->amount !== null) {
-                $numCombinations = count($combinations);
-                $payoutAmount = (int) ($totalPayout * $purchase->amount / ($numCombinations * 100));
+            $payoutAmount = $this->calculatePayoutAmountAction->execute($purchase);
+            if ($payoutAmount !== null) {
                 $purchase->payout_amount = $payoutAmount;
                 $purchase->save();
             }
         }
-    }
-
-    /**
-     * buy_type と ticket_type の指定頭数に応じて selections を組み合わせに展開する。
-     *
-     * 着順考慮（umatan / sanrentan）の場合、各組み合わせは順序を保持した配列。
-     * 着順不問の場合、各組み合わせは昇順ソート済みの配列。
-     *
-     * @param  array<string, mixed>  $selections
-     * @return array<int, array<int, int>>
-     */
-    private function expandSelections(string $buyType, array $selections, int $horseCount, bool $isOrdered): array
-    {
-        $combinations = match ($buyType) {
-            'single' => $this->expandSingle($selections, $horseCount, $isOrdered),
-            'box' => $this->expandBox($selections, $horseCount, $isOrdered),
-            'nagashi' => $this->expandNagashi($selections, $horseCount, $isOrdered),
-            'formation' => $this->expandFormation($selections, $horseCount, $isOrdered),
-            default => [],
-        };
-
-        return $this->normalizeCombinations($combinations, $isOrdered);
-    }
-
-    /**
-     * @param  array<string, mixed>  $selections
-     * @return array<int, array<int, int>>
-     */
-    private function expandSingle(array $selections, int $horseCount, bool $isOrdered): array
-    {
-        $horses = $this->extractIntList($selections['horses'] ?? []);
-
-        if ($horseCount === 1) {
-            return array_map(static fn (int $h): array => [$h], $horses);
-        }
-
-        if (count($horses) !== $horseCount) {
-            return [];
-        }
-
-        return [$horses];
-    }
-
-    /**
-     * @param  array<string, mixed>  $selections
-     * @return array<int, array<int, int>>
-     */
-    private function expandBox(array $selections, int $horseCount, bool $isOrdered): array
-    {
-        $horses = $this->extractIntList($selections['horses'] ?? []);
-
-        if ($horseCount === 1) {
-            return array_map(static fn (int $h): array => [$h], $horses);
-        }
-
-        if (count($horses) < $horseCount) {
-            return [];
-        }
-
-        if ($isOrdered) {
-            return $this->permutations($horses, $horseCount);
-        }
-
-        return $this->combinations($horses, $horseCount);
-    }
-
-    /**
-     * @param  array<string, mixed>  $selections
-     * @return array<int, array<int, int>>
-     */
-    private function expandNagashi(array $selections, int $horseCount, bool $isOrdered): array
-    {
-        $axis = $this->extractIntList($selections['axis'] ?? []);
-        $others = $this->extractIntList($selections['others'] ?? []);
-
-        if ($horseCount === 1) {
-            return array_map(
-                static fn (int $h): array => [$h],
-                array_values(array_unique(array_merge($axis, $others)))
-            );
-        }
-
-        if ($axis === [] || $others === []) {
-            return [];
-        }
-
-        $axisNeeded = $horseCount - 1;
-        if (count($axis) < $axisNeeded || count($axis) > $horseCount - 1) {
-            // axis が 1 頭または 2 頭のパターンに対応するが、標準は axisNeeded = horseCount - 1。
-            // 3連複で軸2頭の場合は axisNeeded = 2 に該当する。
-        }
-
-        // 軸の組み合わせ数 = horseCount - 1（相手が 1 頭）か、軸の頭数による。
-        // nagashi の一般化: 軸は axis 全件をそのまま含め、残り (horseCount - count(axis)) 頭を others から選ぶ。
-        $remainingSlots = $horseCount - count($axis);
-        if ($remainingSlots < 1 || count($others) < $remainingSlots) {
-            return [];
-        }
-
-        $otherCombinations = $this->combinations($others, $remainingSlots);
-
-        $result = [];
-        foreach ($otherCombinations as $otherCombo) {
-            $combined = array_merge($axis, $otherCombo);
-            if ($isOrdered) {
-                // 着順考慮の nagashi は複数の着順パターンを生成する必要があるが、
-                // 本実装では軸の着順を保持した配列として扱う。
-                $result[] = $combined;
-            } else {
-                $result[] = $combined;
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * @param  array<string, mixed>  $selections
-     * @return array<int, array<int, int>>
-     */
-    private function expandFormation(array $selections, int $horseCount, bool $isOrdered): array
-    {
-        $columns = $selections['columns'] ?? [];
-        if (! is_array($columns) || count($columns) !== $horseCount) {
-            return [];
-        }
-
-        /** @var array<int, array<int, int>> $columnLists */
-        $columnLists = [];
-        foreach ($columns as $column) {
-            $list = $this->extractIntList(is_array($column) ? $column : []);
-            if ($list === []) {
-                return [];
-            }
-            $columnLists[] = $list;
-        }
-
-        $result = [[]];
-        foreach ($columnLists as $column) {
-            $next = [];
-            foreach ($result as $partial) {
-                foreach ($column as $horse) {
-                    if (in_array($horse, $partial, true)) {
-                        continue;
-                    }
-                    $next[] = [...$partial, $horse];
-                }
-            }
-            $result = $next;
-        }
-
-        return $result;
-    }
-
-    /**
-     * 組み合わせ配列を券種の性質に応じて正規化し、重複を除去する。
-     *
-     * @param  array<int, array<int, int>>  $combinations
-     * @return array<int, array<int, int>>
-     */
-    private function normalizeCombinations(array $combinations, bool $isOrdered): array
-    {
-        $seen = [];
-        $normalized = [];
-        foreach ($combinations as $combo) {
-            $key = $isOrdered ? implode('-', $combo) : implode('-', $this->sortedCopy($combo));
-            if (isset($seen[$key])) {
-                continue;
-            }
-            $seen[$key] = true;
-            $normalized[] = $isOrdered ? $combo : $this->sortedCopy($combo);
-        }
-
-        return $normalized;
-    }
-
-    /**
-     * 展開済みの組み合わせと払い戻し結果を照合し、ヒットした金額を合算する。
-     *
-     * @param  array<int, array<int, int>>  $combinations
-     * @param  array<int, array{amount: int, horse_numbers: array<int, int>}>  $payouts
-     */
-    private function sumMatchedPayouts(array $combinations, array $payouts, bool $isOrdered): int
-    {
-        if ($payouts === []) {
-            return 0;
-        }
-
-        $total = 0;
-        foreach ($payouts as $payout) {
-            $target = $isOrdered ? $payout['horse_numbers'] : $this->sortedCopy($payout['horse_numbers']);
-            foreach ($combinations as $combo) {
-                if ($combo === $target) {
-                    $total += $payout['amount'];
-                    break;
-                }
-            }
-        }
-
-        return $total;
-    }
-
-    /**
-     * @param  mixed  $value
-     * @return array<int, int>
-     */
-    private function extractIntList(mixed $value): array
-    {
-        if (! is_array($value)) {
-            return [];
-        }
-
-        $result = [];
-        foreach ($value as $item) {
-            if (is_int($item)) {
-                $result[] = $item;
-            } elseif (is_string($item) && ctype_digit($item)) {
-                $result[] = (int) $item;
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * nCk の組み合わせを生成する（順序不問）。
-     *
-     * @param  array<int, int>  $items
-     * @return array<int, array<int, int>>
-     */
-    private function combinations(array $items, int $k): array
-    {
-        if ($k === 0) {
-            return [[]];
-        }
-        if ($k > count($items)) {
-            return [];
-        }
-
-        $result = [];
-        $length = count($items);
-        for ($i = 0; $i <= $length - $k; $i++) {
-            $head = $items[$i];
-            $tailCombinations = $this->combinations(array_slice($items, $i + 1), $k - 1);
-            foreach ($tailCombinations as $tail) {
-                $result[] = [$head, ...$tail];
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * nPk の順列を生成する（順序あり）。
-     *
-     * @param  array<int, int>  $items
-     * @return array<int, array<int, int>>
-     */
-    private function permutations(array $items, int $k): array
-    {
-        if ($k === 0) {
-            return [[]];
-        }
-
-        $result = [];
-        foreach ($items as $index => $item) {
-            $rest = $items;
-            array_splice($rest, $index, 1);
-            foreach ($this->permutations($rest, $k - 1) as $tail) {
-                $result[] = [$item, ...$tail];
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * @param  array<int, int>  $items
-     * @return array<int, int>
-     */
-    private function sortedCopy(array $items): array
-    {
-        $copy = $items;
-        sort($copy);
-
-        return $copy;
     }
 
     /**
