+++ source/app/UseCases/TicketPurchase/CalculatePayoutAmountAction.php
@@ -1,0 +1,360 @@
+<?php
+
+namespace App\UseCases\TicketPurchase;
+
+use App\Models\RacePayout;
+use App\Models\TicketPurchase;
+
+/**
+ * TicketPurchase の払い戻し金額を計算する。
+ *
+ * race_payouts が存在しない場合・amount が null の場合は null を返す。
+ * 計算式: ヒット組み合わせの払戻合計（JRA公表・100円あたり）× (amount / (組み合わせ数 × 100))
+ */
+class CalculatePayoutAmountAction
+{
+    /** 着順を保持する券種 */
+    private const ORDERED_TYPES = ['umatan', 'sanrentan'];
+
+    /** 券種ごとの照合対象馬番数 */
+    private const TICKET_TYPE_HORSE_COUNT = [
+        'tansho' => 1,
+        'fukusho' => 1,
+        'wakuren' => 2,
+        'umaren' => 2,
+        'umatan' => 2,
+        'wide' => 2,
+        'sanrenpuku' => 3,
+        'sanrentan' => 3,
+    ];
+
+    /**
+     * 対象の TicketPurchase の払い戻し金額を計算して返す。
+     * race_payouts が存在しない場合・amount が null の場合は null を返す。
+     */
+    public function execute(TicketPurchase $purchase): ?int
+    {
+        if ($purchase->amount === null) {
+            return null;
+        }
+
+        $purchase->loadMissing(['ticketType', 'buyType']);
+
+        $ticketTypeName = $purchase->ticketType->name;
+        $buyTypeName = $purchase->buyType->name;
+        $selections = $purchase->selections ?? [];
+
+        $horseCount = self::TICKET_TYPE_HORSE_COUNT[$ticketTypeName] ?? null;
+        if ($horseCount === null) {
+            return null;
+        }
+
+        $isOrdered = in_array($ticketTypeName, self::ORDERED_TYPES, true);
+        $combinations = $this->expandSelections($buyTypeName, $selections, $horseCount, $isOrdered);
+
+        if ($combinations === []) {
+            return null;
+        }
+
+        /** @var \Illuminate\Database\Eloquent\Collection<int, RacePayout> $payouts */
+        $payouts = RacePayout::with(['ticketType', 'racePayoutHorses' => function ($query): void {
+            $query->orderBy('sort_order');
+        }])->where('race_id', $purchase->race_id)->get();
+
+        if ($payouts->isEmpty()) {
+            return null;
+        }
+
+        /** @var array<string, array<int, array{amount: int, horse_numbers: array<int, int>}>> $payoutsByType */
+        $payoutsByType = [];
+        foreach ($payouts as $payout) {
+            $typeName = $payout->ticketType->name;
+            $horseNumbers = $payout->racePayoutHorses
+                ->sortBy('sort_order')
+                ->pluck('horse_number')
+                ->map(fn ($n) => (int) $n)
+                ->values()
+                ->all();
+
+            $payoutsByType[$typeName] ??= [];
+            $payoutsByType[$typeName][] = [
+                'amount' => (int) $payout->payout_amount,
+                'horse_numbers' => $horseNumbers,
+            ];
+        }
+
+        $totalPayout = $this->sumMatchedPayouts(
+            $combinations,
+            $payoutsByType[$ticketTypeName] ?? [],
+            $isOrdered,
+        );
+
+        if ($totalPayout <= 0) {
+            return null;
+        }
+
+        $numCombinations = count($combinations);
+
+        return (int) ($totalPayout * $purchase->amount / ($numCombinations * 100));
+    }
+
+    /**
+     * buy_type と ticket_type の指定頭数に応じて selections を組み合わせに展開する。
+     *
+     * @param  array<string, mixed>  $selections
+     * @return array<int, array<int, int>>
+     */
+    private function expandSelections(string $buyType, array $selections, int $horseCount, bool $isOrdered): array
+    {
+        $combinations = match ($buyType) {
+            'single' => $this->expandSingle($selections, $horseCount, $isOrdered),
+            'box' => $this->expandBox($selections, $horseCount, $isOrdered),
+            'nagashi' => $this->expandNagashi($selections, $horseCount, $isOrdered),
+            'formation' => $this->expandFormation($selections, $horseCount, $isOrdered),
+            default => [],
+        };
+
+        return $this->normalizeCombinations($combinations, $isOrdered);
+    }
+
+    /**
+     * @param  array<string, mixed>  $selections
+     * @return array<int, array<int, int>>
+     */
+    private function expandSingle(array $selections, int $horseCount, bool $isOrdered): array
+    {
+        $horses = $this->extractIntList($selections['horses'] ?? []);
+
+        if ($horseCount === 1) {
+            return array_map(static fn (int $h): array => [$h], $horses);
+        }
+
+        if (count($horses) !== $horseCount) {
+            return [];
+        }
+
+        return [$horses];
+    }
+
+    /**
+     * @param  array<string, mixed>  $selections
+     * @return array<int, array<int, int>>
+     */
+    private function expandBox(array $selections, int $horseCount, bool $isOrdered): array
+    {
+        $horses = $this->extractIntList($selections['horses'] ?? []);
+
+        if ($horseCount === 1) {
+            return array_map(static fn (int $h): array => [$h], $horses);
+        }
+
+        if (count($horses) < $horseCount) {
+            return [];
+        }
+
+        if ($isOrdered) {
+            return $this->permutations($horses, $horseCount);
+        }
+
+        return $this->combinations($horses, $horseCount);
+    }
+
+    /**
+     * @param  array<string, mixed>  $selections
+     * @return array<int, array<int, int>>
+     */
+    private function expandNagashi(array $selections, int $horseCount, bool $isOrdered): array
+    {
+        $axis = $this->extractIntList($selections['axis'] ?? []);
+        $others = $this->extractIntList($selections['others'] ?? []);
+
+        if ($horseCount === 1) {
+            return array_map(
+                static fn (int $h): array => [$h],
+                array_values(array_unique(array_merge($axis, $others)))
+            );
+        }
+
+        if ($axis === [] || $others === []) {
+            return [];
+        }
+
+        $remainingSlots = $horseCount - count($axis);
+        if ($remainingSlots < 1 || count($others) < $remainingSlots) {
+            return [];
+        }
+
+        $otherCombinations = $this->combinations($others, $remainingSlots);
+
+        $result = [];
+        foreach ($otherCombinations as $otherCombo) {
+            $result[] = array_merge($axis, $otherCombo);
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  array<string, mixed>  $selections
+     * @return array<int, array<int, int>>
+     */
+    private function expandFormation(array $selections, int $horseCount, bool $isOrdered): array
+    {
+        $columns = $selections['columns'] ?? [];
+        if (! is_array($columns) || count($columns) !== $horseCount) {
+            return [];
+        }
+
+        /** @var array<int, array<int, int>> $columnLists */
+        $columnLists = [];
+        foreach ($columns as $column) {
+            $list = $this->extractIntList(is_array($column) ? $column : []);
+            if ($list === []) {
+                return [];
+            }
+            $columnLists[] = $list;
+        }
+
+        $result = [[]];
+        foreach ($columnLists as $column) {
+            $next = [];
+            foreach ($result as $partial) {
+                foreach ($column as $horse) {
+                    if (in_array($horse, $partial, true)) {
+                        continue;
+                    }
+                    $next[] = [...$partial, $horse];
+                }
+            }
+            $result = $next;
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  array<int, array<int, int>>  $combinations
+     * @return array<int, array<int, int>>
+     */
+    private function normalizeCombinations(array $combinations, bool $isOrdered): array
+    {
+        $seen = [];
+        $normalized = [];
+        foreach ($combinations as $combo) {
+            $key = $isOrdered ? implode('-', $combo) : implode('-', $this->sortedCopy($combo));
+            if (isset($seen[$key])) {
+                continue;
+            }
+            $seen[$key] = true;
+            $normalized[] = $isOrdered ? $combo : $this->sortedCopy($combo);
+        }
+
+        return $normalized;
+    }
+
+    /**
+     * @param  array<int, array<int, int>>  $combinations
+     * @param  array<int, array{amount: int, horse_numbers: array<int, int>}>  $payouts
+     */
+    private function sumMatchedPayouts(array $combinations, array $payouts, bool $isOrdered): int
+    {
+        if ($payouts === []) {
+            return 0;
+        }
+
+        $total = 0;
+        foreach ($payouts as $payout) {
+            $target = $isOrdered ? $payout['horse_numbers'] : $this->sortedCopy($payout['horse_numbers']);
+            foreach ($combinations as $combo) {
+                if ($combo === $target) {
+                    $total += $payout['amount'];
+                    break;
+                }
+            }
+        }
+
+        return $total;
+    }
+
+    /**
+     * @param  mixed  $value
+     * @return array<int, int>
+     */
+    private function extractIntList(mixed $value): array
+    {
+        if (! is_array($value)) {
+            return [];
+        }
+
+        $result = [];
+        foreach ($value as $item) {
+            if (is_int($item)) {
+                $result[] = $item;
+            } elseif (is_string($item) && ctype_digit($item)) {
+                $result[] = (int) $item;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  array<int, int>  $items
+     * @return array<int, array<int, int>>
+     */
+    private function combinations(array $items, int $k): array
+    {
+        if ($k === 0) {
+            return [[]];
+        }
+        if ($k > count($items)) {
+            return [];
+        }
+
+        $result = [];
+        $length = count($items);
+        for ($i = 0; $i <= $length - $k; $i++) {
+            $head = $items[$i];
+            $tailCombinations = $this->combinations(array_slice($items, $i + 1), $k - 1);
+            foreach ($tailCombinations as $tail) {
+                $result[] = [$head, ...$tail];
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  array<int, int>  $items
+     * @return array<int, array<int, int>>
+     */
+    private function permutations(array $items, int $k): array
+    {
+        if ($k === 0) {
+            return [[]];
+        }
+
+        $result = [];
+        foreach ($items as $index => $item) {
+            $rest = $items;
+            array_splice($rest, $index, 1);
+            foreach ($this->permutations($rest, $k - 1) as $tail) {
+                $result[] = [$item, ...$tail];
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  array<int, int>  $items
+     * @return array<int, int>
+     */
+    private function sortedCopy(array $items): array
+    {
+        $copy = $items;
+        sort($copy);
+
+        return $copy;
+    }
+}
