+++ source/tests/Feature/Tickets/TicketPurchaseTest.php
@@ -56,3 +56,77 @@ test('authenticated user can purchase a ticket', function () {
         'amount' => 100,
     ]);
 });
+
+test('レース結果登録済みの場合、馬券購入時に payout_amount が計算される', function () {
+    // Arrange
+    $user = User::factory()->create();
+
+    $now = now();
+
+    $venueId = DB::table('venues')->insertGetId([
+        'name' => '東京',
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    $tanshoTypeId = DB::table('ticket_types')->insertGetId([
+        'name' => 'tansho',
+        'label' => '単勝',
+        'sort_order' => 1,
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    DB::table('buy_types')->insert([
+        'name' => 'single',
+        'label' => '通常',
+        'sort_order' => 1,
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    $raceId = DB::table('races')->insertGetId([
+        'uid' => 'test-uid-' . uniqid(),
+        'venue_id' => $venueId,
+        'race_date' => '2026-04-05',
+        'race_number' => 1,
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    // race_payouts（単勝3番 610円）を登録済みにする
+    $payoutId = DB::table('race_payouts')->insertGetId([
+        'race_id' => $raceId,
+        'ticket_type_id' => $tanshoTypeId,
+        'payout_amount' => 610,
+        'popularity' => 2,
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    DB::table('race_payout_horses')->insert([
+        'race_payout_id' => $payoutId,
+        'horse_number' => 3,
+        'sort_order' => 1,
+        'created_at' => $now,
+        'updated_at' => $now,
+    ]);
+
+    // Act
+    $this->actingAs($user)->post(route('tickets.store'), [
+        'venue' => '東京',
+        'race_date' => '2026-04-05',
+        'race_number' => 1,
+        'ticket_type' => 'tansho',
+        'buy_type' => 'single',
+        'selections' => ['horses' => [3]],
+        'amount' => 100,
+    ]);
+
+    // Assert
+    $this->assertDatabaseHas('ticket_purchases', [
+        'user_id' => $user->id,
+        'race_id' => $raceId,
+        'payout_amount' => 610,
+    ]);
+});
